requirements.txt
yfinance
pandas
numpy
xgboost
nltk
matplotlib
scikit-learn
joblib
requests
GoogleNews
alpaca-trade-api
